Move miscellaneous implementations into interface packages
