+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/impl/DefaultInstrumentStatusService.java
@@ -1,4 +1,4 @@
-package uk.ac.bbsrc.tgac.miso.service;
+package uk.ac.bbsrc.tgac.miso.service.impl;
 
 import java.io.IOException;
 import java.util.List;
@@ -9,6 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import uk.ac.bbsrc.tgac.miso.core.data.InstrumentStatus;
 import uk.ac.bbsrc.tgac.miso.core.store.InstrumentStatusStore;
+import uk.ac.bbsrc.tgac.miso.service.InstrumentStatusService;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
